refactor(main): extract shard s1 union (tui_launch + bundled_skills_sync) from main.py
